refactor: move middlewares to src
